+++ core/components/rememberthis/lexicon/de/default.inc.php
@@ -5,6 +5,7 @@
  * @package rememberthis
  * @subpackage lexicon
  */
+$_lang['rememberthis'] = 'RememberThis';
 $_lang['rememberthis.add'] = 'Merken';
 $_lang['rememberthis.delete'] = 'Entfernen';
 $_lang['rememberthis.noresultstext'] = 'Die Liste ist leer';