+++ core/components/rememberthis/lexicon/de/properties.inc.php
@@ -7,13 +7,18 @@
  */
 $_lang['rememberthis.rememberthisadd.addId'] = 'Dieser Wert, der dazugehörige Titel und eventuelle Eigenschaften wird der Liste hinzugefügt.';
 $_lang['rememberthis.rememberthisadd.addTpl'] = 'Template für den Hinzufügen Link.';
-$_lang['rememberthis.rememberthishook.clearList'] = 'Liste leeren nachdem der Hook ausgeführt wurde.';
-$_lang['rememberthis.rememberthishook.jsonList'] = 'Ausgabe als JSON kodiertes Array von assoziativen Arrays von Element-Identifiern und zugehörige Element-Eigenschaften.';
+$_lang['rememberthis.rememberthisadd.tplPath'] = 'Basis Pfad für Template Chunks mit @FILE Binding.';
+$_lang['rememberthis.rememberthishook.rememberClearList'] = 'Liste leeren nachdem der Hook ausgeführt wurde.';
+$_lang['rememberthis.rememberthishook.rememberJsonList'] = 'Ausgabe als JSON kodiertes Array von assoziativen Arrays von Element-Identifiern und zugehörige Element-Eigenschaften.';
+$_lang['rememberthis.rememberthishook.rememberNoResultsTpl'] = 'Template das angezeigt wird, wenn die Liste leer ist.';
 $_lang['rememberthis.rememberthishook.rememberOuterTpl'] = 'Außen-Template für die Ausgabe der Liste, wenn die Liste nicht leer ist.';
 $_lang['rememberthis.rememberthishook.rememberRowTpl'] = 'Zeilen-Template für die Ausgabe der Listenelemente.';
-$_lang['rememberthis.rememberthishook.saveList'] = 'Die gemerkte Liste mit einem Hash Wert in der Datenbank abspeichern und einen Platzhalter mit diesem Hash füllen.';
+$_lang['rememberthis.rememberthishook.rememberSaveList'] = 'Die gemerkte Liste mit einem Hash Wert in der Datenbank abspeichern und einen Platzhalter mit diesem Hash füllen.';
+$_lang['rememberthis.rememberthishook.rememberTplPath'] = 'Basis Pfad für Template Chunks mit @FILE Binding.';
+$_lang['rememberthis.rememberthishook.rememberWrapperTpl'] = 'Wrapper-Template für die Ausgabe des Mantels oder die leere Liste.';
 $_lang['rememberthis.rememberthislist.jsonList'] = 'Ausgabe als JSON kodiertes Array von assoziativen Arrays von Element-Identifiern und zugehörige Element-Eigenschaften.';
-$_lang['rememberthis.rememberthislist.noResultsTpl'] = 'Template für die leere Liste.';
+$_lang['rememberthis.rememberthislist.noResultsTpl'] = 'Template das angezeigt wird, wenn die Liste leer ist.';
 $_lang['rememberthis.rememberthislist.outerTpl'] = 'Außen-Template für die Ausgabe der Liste, wenn die Liste nicht leer ist.';
 $_lang['rememberthis.rememberthislist.rowTpl'] = 'Zeilen-Template für die Ausgabe der Listenelemente.';
+$_lang['rememberthis.rememberthislist.tplPath'] = 'Basis Pfad für Template Chunks mit @FILE Binding.';
 $_lang['rememberthis.rememberthislist.wrapperTpl'] = 'Wrapper-Template für die Ausgabe des Mantels oder die leere Liste.';