+++ core/components/rememberthis/lexicon/en/default.inc.php
@@ -5,6 +5,7 @@
  * @package rememberthis
  * @subpackage lexicon
  */
+$_lang['rememberthis'] = 'RememberThis';
 $_lang['rememberthis.add'] = 'Remember';
 $_lang['rememberthis.delete'] = 'Delete';
 $_lang['rememberthis.noresultstext'] = 'The List is empty';