+++ core/components/rememberthis/lexicon/en/properties.inc.php
@@ -7,13 +7,18 @@
  */
 $_lang['rememberthis.rememberthisadd.addId'] = 'This value (and the correspondenting item title) is added to the list.';
 $_lang['rememberthis.rememberthisadd.addTpl'] = 'Template for the add link.';
-$_lang['rememberthis.rememberthishook.clearList'] = 'Clear the list after running the hook.';
-$_lang['rememberthis.rememberthishook.jsonList'] = 'Output a JSON-encoded array of associative arrays of element identifiers and element itemproperties.';
+$_lang['rememberthis.rememberthisadd.tplPath'] = 'Base path for template chunks using @FILE binding.';
+$_lang['rememberthis.rememberthishook.rememberClearList'] = 'Clear the list after running the hook.';
+$_lang['rememberthis.rememberthishook.rememberJsonList'] = 'Output a JSON-encoded array of associative arrays of element identifiers and element itemproperties.';
+$_lang['rememberthis.rememberthishook.rememberNoResultsTpl'] = 'Template that is displayed, if the list is empty.';
 $_lang['rememberthis.rememberthishook.rememberOuterTpl'] = 'Outer template for the output of the list, if the list is not empty.';
 $_lang['rememberthis.rememberthishook.rememberRowTpl'] = 'Row template for the output of a list element.';
-$_lang['rememberthis.rememberthishook.saveList'] = 'Save the remembered list with a hash in the database and fill a placeholder with this hash.';
+$_lang['rememberthis.rememberthishook.rememberSaveList'] = 'Save the remembered list with a hash in the database and fill a placeholder with this hash.';
+$_lang['rememberthis.rememberthishook.rememberTplPath'] = 'Base path for template chunks using @FILE binding.';
+$_lang['rememberthis.rememberthishook.rememberWrapperTpl'] = 'Wrapper template for the outer output or the empty output.';
 $_lang['rememberthis.rememberthislist.jsonList'] = 'Output a JSON-encoded array of associative arrays of element identifiers and element itemproperties.';
 $_lang['rememberthis.rememberthislist.noResultsTpl'] = 'Template that is displayed, if the list is empty.';
 $_lang['rememberthis.rememberthislist.outerTpl'] = 'Outer template for the output of the list, if the list is not empty.';
 $_lang['rememberthis.rememberthislist.rowTpl'] = 'Row template for the output of a list element.';
+$_lang['rememberthis.rememberthislist.tplPath'] = 'Base path for template chunks using @FILE binding.';
 $_lang['rememberthis.rememberthislist.wrapperTpl'] = 'Wrapper template for the outer output or the empty output.';