+++ core/components/rememberthis/src/Snippets/RememberThisClear.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * RememberThis2FormIt Hook
+ *
+ * @package rememberthis
+ * @subpackage hook
+ */
+
+namespace TreehillStudio\RememberThis\Snippets;
+
+class RememberThisClear extends Hook
+{
+    /**
+     * Execute the hook and return success.
+     *
+     * @return bool
+     * @throws /Exception
+     */
+    public function execute()
+    {
+        $this->rememberthis->clearList();
+
+        return true;
+    }
+}